pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "artus"
version = "0.1.0"
authors = [
  { name="Justine Talpaert", email="justine.daudon@protonmail.com" },
]
description = "A geoAI package to train and predict spatial occurences on rasters or georeferenced images."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "chardet==5.1.0", 
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git@072713649a9b0069c10aad1aaab819112e8f1e2e",
    "fiftyone==0.19.1",
    "fiftyone-brain==0.10.0",
    "fiftyone-db==0.4.0",
    "Fiona==1.8.22",
    "GDAL==3.5.3",
    "importlib-metadata==6.1.0",
    "importlib-resources==5.7.1",
    "iopath==0.1.9",
    "ipykernel==6.13.0",
    "ipython==8.4.0",
    "ipython-genutils==0.2.0",
    "ipywidgets==7.7.0",
    "jedi==0.18.1",
    "jmespath==1.0.1",
    "joblib==1.2.0",
    "jupyter==1.0.0",
    "jupyter-bbox-widget==0.5.0",
    "jupyter-client==7.3.1",
    "jupyter-console==6.4.3",
    "jupyter-core==4.10.0",
    "jupyterlab-pygments==0.2.2",
    "jupyterlab-widgets==1.1.0",
    "Markdown==3.4.1",
    "MarkupSafe==2.1.1",
    "matplotlib>=3.5.0",
    "matplotlib-inline==0.1.3",
    "mistune==0.8.4",
    "mongoengine==0.24.2",
    "notebook==6.4.11",
    "numpy>=1.18.1",
    "omegaconf==2.3.0",
    "opencv-python==4.6.0",
    "pandas>=1.5.2",
    "Pillow==9.2.0",
    "plotly==5.13.0",
    "psutil==5.9.1",
    "pycocotools==2.0.6",
    "pylabel==0.1.41",
    "pymongo==4.3.3",
    "PyYAML>=5.0.0",
    "rasterio==1.3.3",
    "scikit-learn==1.2.2",
    "scipy==1.9.3",
    "segmentation-models-pytorch==0.3.1",
    "shapely>=1.8.5",
    "tabulate==0.9.0",
    "tensorboard==2.12.0",
    "tensorboard-data-server==0.7.0",
    "tensorboard-plugin-wit==1.8.1",
    "tensorboardX==2.5.1",
    "torch==1.12.1",
    "torchmetrics==0.9.3",
    "torchvision==0.13.1",
    "tqdm==4.65.0",
    "voxel51-eta==0.8.4"
]

[project.urls]
"Homepage" = "https://github.com/6tronl/artus"
"Bug Tracker" = "https://github.com/6tronl/artus/issues"
